requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
plotly
pydeck
langgraph
langchain
langchain-huggingface
fpdf2
statsmodels
python-dotenv
requests
